authentication and data type support
Optional authentication as Bearer Authorization HTTP header was added.

All data from the json was previously considered as float, which would lead to errors if the data retrieved was text. Now a new DataType setting for each value must be set in the ini file. There are two supported types: text and float, anything other than text in the DataType setting will be considered a float (so integers will be cast as floats, but this shouldn't be an issue)